2 commit

+++ src/main/java/repository/ClientsRepository.java
@@ -3,6 +3,6 @@ package repository;
 import clients.entity.Client;
 import org.springframework.data.repository.CrudRepository;
 
-
-public interface ClientsRepository extends CrudRepository<Client, Long> {
+public interface ClientsRepository extends CrudRepository<Client, Integer> {
 }
+
